+++ tests/test_workplan_html.py
@@ -20,16 +20,16 @@
 
 """Tests on export of work plan to HTML."""
 
-from .utils import get_json_file
+from .utils import get_input_file
 
 
 def test_workplan_to_html():
     """Test workplan_to_html function."""
     from tasksched import workplan_to_html
-    workplan = get_json_file('workplan_complete.json')
+    workplan = get_input_file('workplan_complete.yaml')
     html = workplan_to_html(workplan)
     assert html.startswith('<!doctype html>')
 
-    workplan = get_json_file('workplan_complete2.json')
+    workplan = get_input_file('workplan_complete2.yaml')
     html = workplan_to_html(workplan)
     assert html.startswith('<!doctype html>')