+++ ChangeLog.md
@@ -2,6 +2,8 @@
 
 ## Version 0.4.0 (under dev)
 
+- Allow YAML files in input, return work plan in YAML by default.
+- Add workplan option `-j`/`--json`.
 - Add HTML output.
 - Accept task duration as float in input, round it to the next largest integer.
 - Ignore tasks with null or negative duration.